Changed the title of the HTML document from "Js arrow" to "Js Classes".
Changed the code from creating an arrow function to creating a class named "car" with a constructor that takes two parameters, "name" and "year".

Also created a new class named "person" with a constructor that takes two parameters, "name" and "year". The "person" class also has a method called "age" that returns the age of the person.

Finally, created two instances of the "car" class and two instances of the "person" class and logged them to the console. (generated by useblackbox.ai )